requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.2
greenlet==2.0.2
typing_extensions==4.8.0
